minor cleanup (avoid infix ops, move some NodeRsp into Node.scala, NodeOpRsp into NodeOp.scala
